+++ Convertidor/Dtos/Rh/RhTmpRetencionesFjpDto.cs
@@ -10,9 +10,9 @@
         public string NombresApellidos { get; set; }
         public string DescripcionCargo { get; set; }
         public DateTime FechaIngreso { get; set; }
-        public int MontoFjpTrabajador { get; set; }
-        public int MontoFjpPatrono { get; set; }
-        public int MontoTotalRetencion { get; set; }
+        public decimal MontoFjpTrabajador { get; set; }
+        public decimal MontoFjpPatrono { get; set; }
+        public decimal MontoTotalRetencion { get; set; }
         public string FechaNomina { get; set; }
         public string SiglasTipoNomina { get; set; }
         public DateTime FechaDesde { get; set; }
